requirements.txt
asgiref==3.8.1
Django==5.1.7
djangorestframework==3.16.0
psycopg2==2.9.10
PyYAML==6.0.2
sqlparse==0.5.3
